requirements.txt
streamlit>=1.32.0
google-genai>=0.3.0